🔀 Merge development (#9)
* Feature/cloudflare preview (#5)

* fix errot type check

* add different preview and production environment roots

* add AuthError typeGuard

* Feature/cloudflare preview (#6)

* fix errot type check

* add different preview and production environment roots

* add AuthError typeGuard

* make root env public

* add delete user badge

* fix delete user endpoint

* add api to allowed origin

* add * to allowed origin

* remove allowed origin

* fix statuscodes

* :heavy_plus_sign: Add tailwind

* :boom: Setup shadcn ui components

* :construction: Update layout and some pages

---------

+++ src/components/delete-user-badge.tsx
@@ -2,22 +2,52 @@ import { useStore } from '@nanostores/react';
 import { auth0Token } from '../stores/tokenStore';
 import { logoutFlow, type AuthToken } from '../auth0';
 import { useState } from 'react';
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from '@/components/ui/alert-dialog';
+import { Button } from '@/components/ui/button';
 
 const BASE_URL = 'https://api.whu.app/users/v1/';
 
 function DeleteUserBadge() {
   const [message, setMessage] = useState<string>('');
   const authToken: AuthToken = JSON.parse(useStore(auth0Token));
+
   if (useStore(auth0Token)) {
     return (
-      <button
-        onClick={async () => {
-          const message = await deleteOwnUser(authToken);
-          setMessage(message);
-        }}
-      >
-        Delete User
-      </button>
+      <AlertDialog>
+        <AlertDialogTrigger asChild>
+          <Button variant='destructive'>Delete Account</Button>
+        </AlertDialogTrigger>
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
+            <AlertDialogDescription>
+              This action cannot be undone. This will permanently delete your account and remove
+              your data from our servers.
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel>Cancel</AlertDialogCancel>
+            <AlertDialogAction
+              onClick={async () => {
+                const message = await deleteOwnUser(authToken);
+                setMessage(message);
+              }}
+            >
+              Continue
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
     );
   } else {
     return (
